heartbeat_cmd_scheduler/requirements.txt
pymongo
schedule
paho-mqtt